// ui/apps/pmm/src/types/severity.types.ts
// comes from api/management/v1/severity.proto
export enum Severity {
  unspecified = 'SEVERITY_UNSPECIFIED',
  emergency = 'SEVERITY_EMERGENCY',
  alert = 'SEVERITY_ALERT',
  critical = 'SEVERITY_CRITICAL',
  error = 'SEVERITY_ERROR',
  warning = 'SEVERITY_WARNING',
  notice = 'SEVERITY_NOTICE',
  info = 'SEVERITY_INFO',
  debug = 'SEVERITY_DEBUG',
}